+++ apps/app/src/app/@seller/transaction-listing/page/ListingTransactionListPage/TransactionList/index.ts
@@ -1,0 +1,1 @@
+export { TransactionList } from "./TransactionList";
