+++ apps/app/src/app/@seller/transaction-listing/page/TransactionListingListPage/TransactionListingList/TransactionListingList.tsx
@@ -1,16 +1,14 @@
-import type { MarkSuspense } from "@use-pico/client/type";
 import { Container } from "@use-pico/client/ui/container";
-import type { FC } from "react";
+import { type FC, Suspense } from "react";
 import { Item } from "./Item";
 
 export namespace TransactionListingList {
-	export interface Props extends Container.Props, MarkSuspense.Props {
+	export interface Props extends Container.Props {
 		transactionListingIds: string[];
 	}
 }
 
 export const TransactionListingList: FC<TransactionListingList.Props> = ({
-	_suspense,
 	transactionListingIds,
 	ui,
 	...props
@@ -26,12 +24,16 @@ export const TransactionListingList: FC<TransactionListingList.Props> = ({
 		>
 			{transactionListingIds.map((transactionListingId) => {
 				return (
-					<Item
+					<Suspense
 						key={transactionListingId}
-						data-id={transactionListingId}
-						_suspense={_suspense}
-						transactionListingId={transactionListingId}
-					/>
+						fallback={<Item.Fallback />}
+					>
+						<Item
+							data-id={transactionListingId}
+							_suspense={"I know"}
+							transactionListingId={transactionListingId}
+						/>
+					</Suspense>
 				);
 			})}
 		</Container>
