+++ apps/app/src/app/@seller/transaction-listing/page/TransactionListingListPage/index.ts
@@ -1,0 +1,1 @@
+export { TransactionListingListPage } from "./TransactionListingListPage";
